requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
dnspython==2.6.1
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
gTTS==2.5.1
idna==3.6
multidict==6.0.5
pycparser==2.21
pymongo==4.6.2
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
